# backend/docs/swagger.yaml
definitions:
...
  main.LoginResponse:
    properties:
      code:
        type: string
      redirect:
        type: string
    type: object
  main.ProfileCreationBody:
    properties:
      bio:
        type: string
      dateOfBirth:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      nickName:
        type: string
      quests:
        items:
          $ref: '#/definitions/main.QuestsCreation'
        type: array
      skills:
        items:
          $ref: '#/definitions/main.SkillsCreation'
        type: array
    type: object
...
  main.QuestsCreation:
    properties:
      experience:
        type: integer
      name:
        type: string
    type: object
...
  main.SkillsCreation:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  main.UpdateUserBody:
    properties:
      bio:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      nickName:
        type: string
      target:
        type: string
    required:
    - target
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Level-Up app
  title: Level-Up API
  version: "0.1"
paths:
...
  /v1/levelup_api/createProfile:
    post:
      consumes:
      - application/json
      description: Create the player with user's info, skills and quests
      parameters:
      - description: Profile creation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.ProfileCreationBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LoginResponse'
      summary: Create the player
      tags:
      - Player
...
  /v1/levelup_api/updateUser:
    post:
      consumes:
      - application/json
      description: Update the player
      parameters:
      - description: User update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.UpdateUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update the player
      tags:
      - Player
...
swagger: "2.0"